==> RunAnalysis/1_make_quality_dt.Rmd <==
---
title: "HGP analysis"
author: 
- name: Rick Farouni, 
  affiliation:
  - &cruk Department of Human Genetics, McGill University,  Montreal, Canada
date: '`r format(Sys.Date(), "%Y-%B-%d")`'
output:
  html_notebook:
    df_print: paged
    code_folding: show
    toc: yes
    toc_float: 
      collapsed: false
      smooth_scroll: false
---

# Preparations

```{r setup}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file(), fig.width=16)
options(scipen=999)
```

```{r}
library(tidyverse)
```


## Read bas files

```{r read_bas}
read_basfiles <- function(bas_filepath) {
  
  bas_file <- read_tsv(bas_filepath,
                  col_types = cols_only('sample' = 'c', 
                                        'platform' = 'c',        
                                       'average_quality_of_mapped_bases' = col_guess())) 
  return( bas_file)
}
```

### Preprocess table



```{r}
bas_files <- list.files(path = "./data/bas",
                        pattern="*bam.bas",
                        full.names=TRUE)
```

```{r}


bas_dt <- 
  map_dfr(bas_files, read_basfiles ) %>%
  rename(avg_qual="average_quality_of_mapped_bases")
bas_dt
```


```{r}
bas_dt <-
  bas_dt %>%
  group_by(sample) %>%
  mutate(id=1:n()) %>%
  select(-platform)%>%
  spread(id,avg_qual)
bas_dt
```
```{r}
n_not_na <- function(x) {
 sum(!is.na(x))
}
```


```{r}
 bas_dt <- 
  bas_dt %>%
  ungroup()%>% 
    mutate( avg_qual_sd =pmap_dbl(select(., -sample), lift_vd(sd,na.rm =TRUE )),
           avg_qual_mean = rowMeans(select(., -sample),na.rm =TRUE),
           avg_qual_n= pmap_dbl(select(., -sample), lift_vd(n_not_na)),
           avg_qual_min= pmap_dbl(select(., -sample), lift_vd(min, na.rm =TRUE)),
           avg_qual_max= pmap_dbl(select(., -sample), lift_vd(max, na.rm =TRUE))) %>%
  select (sample, avg_qual_n, avg_qual_min,avg_qual_max,avg_qual_mean,avg_qual_sd, everything() ) 
```



```{r}
ped <- 
  read_tsv("./data/metadata/20130606_g1k.ped")%>% 
  rename(sample="Individual ID")
ped 
```


```{r}
bas_dt <- inner_join( ped  %>% 
                       select(sample, Population),
                       bas_dt,
                     by="sample")
bas_dt
```


```{r}
bas_filepath <- "./data/metadata/average_quality_dt.txt"
write_tsv(bas_dt, bas_filepath)
```

Plot 
 

```{r fig.width=20}
ggplot(bas_dt,
       aes(x=avg_qual_min,
           y=avg_qual_max, 
           colour=avg_qual_n )) +
  geom_point(size=2)
 
```


```{r fig.width=20}
ggplot(bas_dt,
       aes(x=avg_qual_mean,
           y=avg_qual_sd, 
           colour=avg_qual_n )) +
  geom_point(size=2)
 
```
